refactor: move yaml reader from main file
